requirements.txt
pyrealsense2
xarm-python-sdk
opencv-python
numpy
requests
python-dotenv
pyyaml
